Reorganize components into feature-based folders.
