+++ frontend/src/components/base/Auth/RegisterForm.vue
