+++ frontend/src/components/base/icons/IconCommunity.vue
