+++ frontend/src/components/base/icons/IconDocumentation.vue
