+++ frontend/src/components/base/icons/IconEcosystem.vue
