+++ frontend/src/components/categories/CategoryUploadForm.vue
