+++ frontend/src/components/gallery/GalleryForm.vue
