+++ frontend/src/components/gallery/ImageDetailsModal.vue
