+++ frontend/src/components/items/ItemsDataTable.vue
