+++ frontend/src/views/TheContactsView.vue
@@ -72,7 +72,7 @@
 </template>
 
 <script setup lang="ts">
-import ContactForm from '@/components/ContactForm.vue';
+import ContactForm from '@/components/shared/ContactForm.vue';
 </script>
 
 <style scoped>
